requirements.txt
langchain
streamlit
python-dotenv
PyPDF2
langchain_google_genai
langchain_community
langchain-huggingface
huggingface_hub


-e .